test(webui): share transcript fake DOM harness
